added changes on postPublish.router.js

+++ src/routes/postPublish.router.js
@@ -2,6 +2,7 @@ const express = require('express');
 const postUsecase = require('../usecases/postPulish.usecase');
 const auth = require('../middleware/auth.middleware');
 const getIdToken = require('../lib/getID')
+const createError = require('http-errors')
 
 const router = express.Router();
 
@@ -46,6 +47,9 @@ router.get('/', async (req, res)=>{
 router.patch('/:id', auth, async(req, res)=>{
     try {
         const {id} = req.params;
+        if(req.body.user){
+            throw createError(423, "It is not possible to change the user.")
+        }
         const postUpdated = await postUsecase.updatePostById(id, req.body);
         res.json({
             success: true,
@@ -64,6 +68,16 @@ router.patch('/:id', auth, async(req, res)=>{
 router.delete('/:id', auth, async (req, res)=>{
     try {
         const {id} = req.params;
+
+        const usertoken = await getIdToken(req, res);
+        const userId = usertoken._id
+        const post = await postUsecase.getPostId(id);
+        const postUserId = post.user._id;
+        if(!postUserId===userId){
+            console.log(userId);
+            console.log(postUserId);
+            throw createError(401, "You can't delete other user post.")
+        }
         const postDelete = await postUsecase.deletePostById(id);
         res.json({
             success: true,
